requirements.txt
langchain
langchain_community
langchain_experimental
langgraph
tiktoken
arxiv
wikipedia
flask
pypdf
cassio
langchain-groq
python-dotenv
langchain-google-genai
duckduckgo-search
tavily-python
-e .
